requirements.txt
altgraph==0.16.1
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
macholib==1.11
pefile==2018.8.8
PyInstaller==3.4
PyQt5==5.11.3
PyQt5-sip==4.19.13
pywin32-ctypes==0.2.0
smmap2==2.0.5
